+++ edit_word.php
@@ -445,9 +445,12 @@ function edit_word_do_form($wid, $text_id, $ord, $fromAnn)
         <tr title="Only change uppercase/lowercase!">
             <td class="td1 right"><b>Edit Term:</b></td>
             <td class="td1">
-                <input <?php echo $scrdir; ?> class="notempty checkoutsidebmp" data_info="Term" type="text" 
-                name="WoText" id="wordfield" value="<?php echo tohtml($term); ?>" maxlength="250" size="35" />
-                <img src="icn/status-busy.png" title="Field must not be empty" alt="Field must not be empty" />
+                <input <?php echo $scrdir; ?> class="notempty checkoutsidebmp" 
+                data_info="Term" type="text" 
+                name="WoText" id="wordfield" 
+                value="<?php echo tohtml($term); ?>" maxlength="250" size="35" />
+                <img src="icn/status-busy.png" title="Field must not be empty" 
+                alt="Field must not be empty" />
             </td>
         </tr>
             <?php print_similar_terms_tabrow(); ?>
@@ -455,7 +458,8 @@ function edit_word_do_form($wid, $text_id, $ord, $fromAnn)
             <td class="td1 right">Translation:</td>
             <td class="td1">
                 <textarea name="WoTranslation" class="setfocus textarea-noreturn checklength checkoutsidebmp" 
-                data_maxlength="500" data_info="Translation" cols="35" rows="3"><?php echo tohtml($transl); ?></textarea>
+                data_maxlength="500" data_info="Translation" cols="35" 
+                rows="3"><?php echo tohtml($transl); ?></textarea>
             </td>
         </tr>
         <tr>
@@ -467,7 +471,9 @@ function edit_word_do_form($wid, $text_id, $ord, $fromAnn)
         <tr>
             <td class="td1 right">Romaniz.:</td>
             <td class="td1">
-                <input type="text" class="checkoutsidebmp" data_info="Romanization" name="WoRomanization" maxlength="100" size="35" 
+                <input type="text" class="checkoutsidebmp" 
+                data_info="Romanization" name="WoRomanization" maxlength="100" 
+                size="35" 
                 value="<?php echo tohtml($record['WoRomanization']); ?>" />
             </td>
         </tr>
@@ -475,8 +481,9 @@ function edit_word_do_form($wid, $text_id, $ord, $fromAnn)
             <td class="td1 right">Sentence<br />Term in {...}:</td>
             <td class="td1">
                 <textarea <?php echo $scrdir; ?> name="WoSentence" 
-                class="textarea-noreturn checklength checkoutsidebmp" data_maxlength="1000" 
-                data_info="Sentence" cols="35" rows="3"><?php echo tohtml($sentence); ?></textarea>
+                class="textarea-noreturn checklength checkoutsidebmp" 
+                data_maxlength="1000" data_info="Sentence" cols="35" 
+                rows="3"><?php echo tohtml($sentence); ?></textarea>
             </td>
         </tr>
         <tr>
@@ -490,11 +497,15 @@ function edit_word_do_form($wid, $text_id, $ord, $fromAnn)
                 <?php 
                 if ($fromAnn !== '') { 
                     echo createDictLinksInEditWin2(
-                        $lang, 'document.forms[0].WoSentence', 'document.forms[0].WoText'
+                        $lang, 
+                        'document.forms[0].WoSentence', 
+                        'document.forms[0].WoText'
                     );
                 } else {
                     echo createDictLinksInEditWin(
-                        $lang, $term, 'document.forms[0].WoSentence', isset($_GET['nodict']) ? 0 : 1
+                        $lang, $term, 
+                        'document.forms[0].WoSentence', 
+                        isset($_GET['nodict']) ? 0 : 1
                     );
                 } 
                 ?>